+++ conformance/src/test/scala/zio/openfeature/conformance/ConformanceSteps.scala
@@ -176,6 +176,7 @@ class ConformanceSteps extends ScalaDsl with EN {
     case ResolutionReason.Unknown        => "UNKNOWN"
     case ResolutionReason.Stale          => "STALE"
     case ResolutionReason.Error          => "ERROR"
+    case ResolutionReason.Other(v)       => v
   }
 
   private def errorCodeName(c: ErrorCode): String = c match {